chore(node): move Node catchup logic to catchup module
